backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
openai-whisper==20240930
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
tqdm==4.66.1
